Move Driver/Provider location options to specific namespace
