+++ DubUrl.Core/Locating/OdbcDriver/Implementation/TimescaleDriverLocator.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
+using DubUrl.Locating.Options;
 
 namespace DubUrl.Locating.OdbcDriver.Implementation
 {
