+++ DubUrl.Core/Locating/Options/EncodingOption.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace DubUrl.Locating.OdbcDriver
+namespace DubUrl.Locating.Options
 {
-    [DriverLocatorOption]
+    [LocatorOption]
     public enum EncodingOption
     {
         Unspecified,
